src/django_template/urls.py
from django.apps import apps
from django.conf import settings
from django.contrib import admin
from django.urls import include, path

from .app.views import HomeView

urlpatterns = [
    path("admin/", admin.site.urls),
    path("", HomeView.as_view(), name="root"),
]

if settings.DEBUG and apps.is_installed("debug_toolbar"):
    import debug_toolbar

    urlpatterns = [
        path("__debug__/", include(debug_toolbar.urls)),
    ] + urlpatterns
