requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/base.in
#
asgiref==3.5.0
    # via django
black==22.1.0
    # via -r requirements/base.in
click==8.0.4
    # via
    #   black
    #   pip-tools
coverage==6.3.2
    # via -r requirements/base.in
django==3.2.11
    # via
    #   -r requirements/base.in
    #   django-debug-toolbar
django-debug-toolbar==3.2.4
    # via -r requirements/base.in
gunicorn==20.1.0
    # via -r requirements/base.in
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pip-tools==6.5.1
    # via -r requirements/base.in
platformdirs==2.5.1
    # via black
psycopg2-binary==2.9.3
    # via -r requirements/base.in
python-dotenv==0.19.2
    # via -r requirements/base.in
pytz==2021.3
    # via django
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
tomli==2.0.1
    # via black
typing-extensions==4.1.1
    # via black
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
